+++ core/php/AbeilleCmd.php
@@ -64,13 +64,35 @@
             logMessage('debug', "  updateDeviceFromDB() WARNING: Unknown addr ${addr} and IEEE is undefined");
             return;
         }
-        if (!isset($GLOBALS['devices'][$net][$addr])) {
-            // Address may have changed following new device announce
+
+        $found = false;
+        if (isset($GLOBALS['devices'][$net][$addr]))
+            $found = true;
+        if ($found == false) {
+            // Address may have changed following new 'device announce'
             foreach ($GLOBALS['devices'][$net] as $addr2 => $eq2) {
                 if ($eq2['ieee'] == $ieee) {
+                    $found = true;
                     $GLOBALS['devices'][$net][$addr] = $GLOBALS['devices'][$net][$addr2];
                     unset($GLOBALS['devices'][$net][$addr2]);
                     logMessage('debug', "  Device ID ${eqId} address changed from ${addr2} to ${addr}.");
+                    break;
+                }
+            }
+        }
+        if ($found == false) {
+            // Equipment may have been migrated from another network
+            foreach ($GLOBALS['devices'] as $net2 => $devices2) { // Go thru all networks
+                if ($net2 == $net)
+                    continue; // This network has already been checked
+                foreach ($GLOBALS['devices'][$net2] as $addr2 => $eq2) {
+                    if ($eq2['ieee'] == $ieee) {
+                        $found = true;
+                        $GLOBALS['devices'][$net][$addr] = $GLOBALS['devices'][$net2][$addr2];
+                        unset($GLOBALS['devices'][$net2][$addr2]);
+                        logMessage('debug', "  Device ID ${eqId} migrated from ${net2}/${addr2} to ${net2}/${addr}.");
+                        break;
+                    }
                 }
             }
         }